iproute2: fix build on GCC 14
The first two patches are already upstream. The third patch is
pending but just adds missing includes. The last patch was not
submitted upstream but just adds a missing include for basename.

The 1st, 3rd and 4th patch add missing includes that would otherwise
break the build with GCC 14 on MUSL. The 2nd patch fixes a newly
raised warning when calling basename with an immutable string.

Signed-off-by: Rany Hany <rany_hany@riseup.net>